Moved core.py to format_manager.py.
