+++ ipymd/__init__.py
@@ -1,3 +1,3 @@
 from . import formats
-from .core.core import convert, format_manager
+from .core.format_manager import convert, format_manager
 from .core.contents_manager import *
