+++ ipymd/core/format_manager.py
