+++ ipymd/formats/tests/test_notebook.py
@@ -6,7 +6,7 @@
 # Imports
 #------------------------------------------------------------------------------
 
-from ...core.core import format_manager, convert
+from ...core.format_manager import format_manager, convert
 from ..notebook import _compare_notebooks
 from ._utils import (_test_reader, _test_writer, _diff, _show_outputs,
                      _exec_test_file, _read_test_file)
