+++ ipymd/formats/tests/test_python.py
@@ -6,7 +6,7 @@
 # Imports
 #------------------------------------------------------------------------------
 
-from ...core.core import format_manager, convert
+from ...core.format_manager import format_manager, convert
 from ...utils.utils import _remove_output
 from ._utils import (_test_reader, _test_writer, _diff, _show_outputs,
                      _exec_test_file, _read_test_file)
